+++ test/anim_search_bar_test.dart
@@ -1,6 +1,6 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:anim_search_bar/anim_search_bar.dart';
+// import 'package:flutter_test/flutter_test.dart';
+//
+// import 'package:anim_search_bar/anim_search_bar.dart';
 //
 // void main() {
 //   test('adds one to input values', () {